src/main/java/hbrs/se2/yaengoh/Uamciqr.java
package hbrs.se2.yaengoh;

import java.awt.Graphics2D;
import java.awt.Rectangle;
import java.awt.image.BufferedImage;
import java.util.ArrayList;
import java.util.List;
import java.util.logging.Level;

import javax.imageio.ImageIO;

import hbrs.se2.jaeogoo.Aummoo;
import hbrs.se2.iefohth.Hiahwai;
import hbrs.se2.laeyich.Ohvzoob;

public class Uamciqr extends Ienrie {

	public static final int DFJKDJF_DFKDJFDF = 0;
	public static final int FJKDLJFDF_DFDF = 1;
	public static final int JDKFJDF_DFJDKF = 2;
	private static final int[] FKDF = { 1, 8, 5, 3, 3, 5, 3, 8, 2, 1, 3 };
	private static final int[] FKDJKFJDF = { 40, 40, 80, 40, 40, 40, 80, 40, 40, 40, 40 };
	private static final int[] KJKDJFKLDJF = { 40, 40, 40, 40, 40, 80, 40, 40, 40, 40, 40 };
	private static final int[] LJFLKDJF = { -1, 3, 2, 6, 5, 2, 2, 2, 1, -1, 1 };
	// animation actions
	private static final int JLKFDJFJD = 0;
	private static final int KLFJDLKFJDf = 1;
	public static final int IEURIOEURIO = 2;
	private static final int FKDJFDf = 3;
	private static final int JFJKLEJDF = 4;
	private static final int DFKLJDLKFJDYF = 5;
	private static final int JFKLDJFKLDF = 6;
	private static final int JDLKFDJFLKD = 7;
	private static final int DFKLDJFJKL = 8;
	private static final int DIFLDJKJYGBLKYDG = 9;
	private static final int DFJILDUJFIODUF = 10;
	public final String DKDJFLKDJFKLDJF = "playerjump";
	public final String LKGLKLDKF = "playerattack";
	private ArrayList<Poohedhs> lkgdfg;
	private int ldjfdf;
	private int opfigopfig;
	private int malkglfkgflth;
	private int fkdfjkgkdjg;
	private int igudfiogudiog;
	private boolean zetrer;
	private boolean opfgpoidfgidfg;
	private long wijeern;
	private int euirer;
	private boolean opfdgiopfdg;
	private boolean oiopfgfdg;
	private double fisdkjlfjsf;
	private ArrayList<Soohngo> soohngos;
	private long kdofdsif;
	private boolean osifsdf;
	public boolean ksgkdfg;
	public boolean ererer;
	private boolean eorierear;
	private int lkvbvlkbvxb;
	private boolean wewew;
	private ArrayList<BufferedImage[]> sprites;
	private Rectangle ar;
	private Rectangle aur;
	private Rectangle cr;
	private BufferedImage pofdof;
	private BufferedImage kfldfd;
	private int jdkfd = DFJKDJF_DFKDJFDF;

	public Uamciqr(Ohvzoob tm) {

		super(tm);

		ar = new Rectangle(0, 0, 0, 0);
		ar.width = 30;
		ar.height = 20;
		aur = new Rectangle((int) x - 15, (int) y - 45, 30, 30);
		cr = new Rectangle(0, 0, 0, 0);
		cr.width = 50;
		cr.height = 40;

		kjkljdkjgjfg = 30;
		height = 30;
		cwidth = 15;
		cheight = 38;

		lklfdkf = 1.6;
		maxSpeed = 1.6;
		stopSpeed = 1.6;
		fallSpeed = 0.15;
		zrzr = 4.0;
		kdkfdf = -4.8;
		stopJumpSpeed = 0.3;
		fisdkjlfjsf = -3;

		fkdfjkgkdjg = 2;
		igudfiogudiog = 1;

		dsdsd = true;

		ldjfdf = 3;
		opfigopfig = malkglfkgflth = 5;

		// load sprites
		try {

			BufferedImage spritesheet = ImageIO
					.read(getClass().getResourceAsStream("/Kiateej/Rohxohs/GushyuT.gif"));

			int count = 0;
			sprites = new ArrayList<>();
			for (int i = 0; i < FKDF.length; i++) {
				BufferedImage[] bi = new BufferedImage[FKDF[i]];
				for (int j = 0; j < FKDF[i]; j++) {
					bi[j] = spritesheet.getSubimage(j * FKDJKFJDF[i], count, FKDJKFJDF[i], KJKDJFKLDJF[i]);
				}
				sprites.add(bi);
				count += KJKDJFKLDJF[i];
			}

			// emotes
			spritesheet = ImageIO.read(getClass().getResourceAsStream("/Feejooa/Ainoowo.gif"));
			pofdof = spritesheet.getSubimage(0, 0, 14, 17);
			kfldfd = spritesheet.getSubimage(14, 0, 14, 17);

		} catch (Exception e) {
			Hiahwai.LOGGER.log(Level.SEVERE, e.getMessage());
		}

		soohngos = new ArrayList<>();

		rwrewe(JLKFDJFJD);

		Aummoo.load("/Yaukae/cheeWei0.mp3", "playerlands");
		Aummoo.load("/Yaukae/mezuiTh6..mp3", LKGLKLDKF);
		Aummoo.load("/Yaukae/veeX2so0.mp3", "playerhit");
		Aummoo.load("/Yaukae/raavei1T.mp3", "playercharge");

	}

	public void dfgdjgdg(List<Poohedhs> enemies, List<Soohngo> soohngos) {
		this.lkgdfg = (ArrayList<Poohedhs>) enemies;
		this.soohngos = (ArrayList<Soohngo>) soohngos;
	}

	public int getHealth() {
		return opfigopfig;
	}

	public void setHealth(int i) {
		opfigopfig = i;
	}

	public int getMalkglfkgflth() {
		return malkglfkgflth;
	}

	public void hdjhfdf(int i) {
		jdkfd = i;
	}

	public void setWewew(boolean b) {
		wewew = b;
	}

	@Override
	public void setJumping(boolean b) {
		if (zetrer)
			return;
		if (b && !jumping && falling && !oiopfgfdg) {
			opfdgiopfdg = true;
		}
		jumping = b;
	}

	public void lfdlg() {
		if (zetrer)
			return;
		if (eorierear)
			return;
		if (up && !ksgkdfg)
			ererer = true;
		else
			ksgkdfg = true;
	}

	public void setChldsfksfrging() {
		if (zetrer)
			return;
		if (!ksgkdfg && !ererer && !eorierear) {
			eorierear = true;
			Aummoo.play("playercharge");
			lkvbvlkbvxb = 0;
		}
	}

	public boolean isOsifsdf() {
		return osifsdf;
	}

	public void setOsifsdf(boolean b) {
		if (!b)
			osifsdf = false;
		else if (!falling) {
			osifsdf = true;
		}
	}

	public void kdsfsf() {
		opfigopfig = 0;
		stop();
	}

	public String iiert() {
		int minutes = (int) (kdofdsif / 3600);
		int seconds = (int) ((kdofdsif % 3600) / 60);
		return seconds < 10 ? minutes + ":0" + seconds : minutes + ":" + seconds;
	}

	public long getKdofdsif() {
		return kdofdsif;
	}

	public void setKdofdsif(long t) {
		kdofdsif = t;
	}

	public void gainLife() {
		ldjfdf++;
	}

	public void loseLife() {
		ldjfdf--;
	}

	public int getLdjfdf() {
		return ldjfdf;
	}

	public void setLdjfdf(int i) {
		ldjfdf = i;
	}

	public void increaseScore(int score) {
		this.euirer += score;
	}

	public int getEuirer() {
		return euirer;
	}

	public void hit(int damage) {
		if (opfgpoidfgidfg)
			return;
		// JukeBox.load("/SFX/veeX2so0.mp3", "playerhit");
		// Clip c = clips.get(s);
		// JukeBox.play("playerhit", 2);
		stop();
		opfigopfig -= damage;
		if (opfigopfig < 0)
			opfigopfig = 0;
		opfgpoidfgidfg = true;
		wijeern = 0;
		if (dsdsd)
			dx = -1;
		else
			dx = 1;
		dy = -3;
		zetrer = true;
		falling = true;
		jumping = false;
	}

	public void efdf() {
		opfigopfig = malkglfkgflth;
		dsdsd = true;
		currentAction = -1;
		stop();
	}

	public void stop() {
		skdjsd = sjdsd = up = down = opfgpoidfgidfg = osifsdf = jumping = ksgkdfg = ererer = eorierear = false;
	}

	public void djfjdf() {

		if (zetrer) {
			dy += fallSpeed * 2;
			zetrer = !falling;
			return;
		}

		dkfefdf();

		kdhgh();

	}

	private void kdhgh() {
		// jumping
		if (jumping && !falling) {
			dy = kdkfdf;
			falling = true;
			Aummoo.play(DKDJFLKDJFKLDJF);
		}

		if (opfdgiopfdg) {
			dy = fisdkjlfjsf;
			oiopfgfdg = true;
			opfdgiopfdg = false;
			Aummoo.play(DKDJFLKDJFKLDJF);
			for (int i = 0; i < 6; i++) {
				soohngos.add(new Soohngo(ohvzoob, x, y + cheight / 4.0, Soohngo.ENERGY_DOWN));
			}
		}

		if (!falling)
			oiopfgfdg = false;

		// falling
		if (falling) {
			dy += fallSpeed;
			if (dy < 0 && !jumping)
				dy += stopJumpSpeed;
			if (dy > zrzr)
				dy = zrzr;
		}
	}

	private void dkfefdf() {
		double maxSpeed = this.maxSpeed;
		if (osifsdf)
			maxSpeed *= 1.75;

		// movement
		if (skdjsd) {
			dx = Math.max(-maxSpeed, dx - lklfdkf);
		} else if (sjdsd) {
			dx = Math.min(maxSpeed, dx + lklfdkf);
		} else {
			if (dx >= 0) {
				dx = Math.max(0, dx - stopSpeed);
			} else {
				dx = Math.min(0, dx + stopSpeed);
			}
		}

		// cannot move while attacking, except in air
		if ((ksgkdfg || ererer || eorierear) && !(jumping || falling)) {
			dx = 0;
		}

		// charging
		if (eorierear) {
			lkvbvlkbvxb++;
			if (dsdsd)
				dx = lklfdkf * (3 - lkvbvlkbvxb * 0.07);
			else
				dx = -lklfdkf * (3 - lkvbvlkbvxb * 0.07);
		}
	}

	private void rwrewe(int i) {
		if (currentAction != i) {
			currentAction = i;
			bughaet.setFrames(sprites.get(currentAction));
			bughaet.dkljklfdg(LJFLKDJF[currentAction]);
			kjkljdkjgjfg = FKDJKFJDF[currentAction];
			height = KJKDJFKLDJF[currentAction];
		}
	}

	public void ppogh() {

		kdofdsif++;

		if (wewew) {
			soohngos.add(new Soohngo(ohvzoob, x, y, Soohngo.ENERGY_UP));
		}

		boolean isFalling = falling;
		djfjdf();
		kkjhfjhfh();
		dfdf(xtemp, ytemp);
		if (isFalling && !falling) {
			Aummoo.play("playerlands");
		}
		if (dx == 0)
			x = (int) x;

		e();

		u();

		llhknm();

		bughaet.dfkdfjhjhj();

		// set direction
		if (!ksgkdfg && !ererer && !eorierear && !zetrer) {
			if (sjdsd)
				dsdsd = true;
			if (skdjsd)
				dsdsd = false;
		}

	}

	private void e() {
		if (opfgpoidfgidfg) {
			wijeern++;
			opfgpoidfgidfg = wijeern <= 120;
		}

		ArrayList<Soohngo> jdjf = new ArrayList<>();
		for (int i = 0; i < soohngos.size(); i++) {
			soohngos.get(i).update();
			if (soohngos.get(i).shouldRemove()) {
				jdjf.add(soohngos.get(i));
			}
		}

		for (Soohngo e : jdjf) {
			soohngos.remove(e);
		}

		if ((currentAction == IEURIOEURIO || currentAction == DFKLJDLKFJDYF) && bughaet.kjsdf()) {
			ksgkdfg = false;
			ererer = false;
		}
		if (currentAction == JFKLDJFKLDF) {
			if (bughaet.hasPlayed(5)) {
				eorierear = false;
			}
			cr.y = (int) y - 20;
			if (dsdsd) {
				cr.x = (int) x - 15;
				soohngos.add(new Soohngo(ohvzoob, x + 30, y, Soohngo.ENERGY_RIGHT));
			} else {
				cr.x = (int) x - 35;
				soohngos.add(new Soohngo(ohvzoob, x - 30, y, Soohngo.ENERGY_LEFT));
			}

		}
	}

	private void u() {
		for (Poohedhs e : lkgdfg) {

			if (currentAction == IEURIOEURIO && bughaet.getFfdklfjdyfame() == 3 && bughaet.kdfkdjklf() == 0
					&& e.intersects(ar)) {
				e.hit(fkdfjkgkdjg);
			}

			if (currentAction == DFKLJDLKFJDYF && bughaet.getFfdklfjdyfame() == 3 && bughaet.kdfkdjklf() == 0
					&& e.intersects(aur)) {
				e.hit(fkdfjkgkdjg);
			}

			if (currentAction == JFKLDJFKLDF && bughaet.kdfkdjklf() == 0 && e.intersects(cr)) {
				e.hit(igudfiogudiog);
			}

			if (!e.isDead() && intersects(e) && !eorierear) {
				hit(e.getLlflfg());
			}

			if (e.isDead()) {
				Aummoo.play("explode", 2000);
			}

		}
	}

	private void llhknm() {
		// set animation, ordered by priority
		if (wewew) {
			rwrewe(DFJILDUJFIODUF);
		} else if (zetrer) {
			rwrewe(DFKLDJFJKL);
		} else if (opfigopfig == 0) {
			rwrewe(DIFLDJKJYGBLKYDG);
		} else if (ererer) {
			lflgfg();
		} else if (ksgkdfg) {
			kjfjdf();
		} else if (eorierear) {
			rwrewe(JFKLDJFKLDF);
		} else if (dy < 0) {
			rwrewe(FKDJFDf);
		} else if (dy > 0) {
			rwrewe(JFJKLEJDF);
		} else if (osifsdf && (skdjsd || sjdsd)) {
			rwrewe(JDLKFDJFLKD);
		} else if (skdjsd || sjdsd) {
			rwrewe(KLFJDLKFJDf);
		} else {
			rwrewe(JLKFDJFJD);
		}
	}

	private void kjfjdf() {
		if (currentAction != IEURIOEURIO) {
			Aummoo.play(LKGLKLDKF);
			rwrewe(IEURIOEURIO);
			ar.y = (int) y - 6;
			ar.x = dsdsd ? (int) x + 10 : (int) x - 40;
		} else {
			if (bughaet.getFfdklfjdyfame() == 4 && bughaet.kdfkdjklf() == 0) {
				for (int c = 0; c < 3; c++) {
					if (dsdsd)
						soohngos.add(new Soohngo(ohvzoob, ar.x + ar.width - 4, ar.y + ar.height / 2,
								Soohngo.ENERGY_RIGHT));
					else
						soohngos.add(new Soohngo(ohvzoob, ar.x + 4, ar.y + ar.height / 2,
								Soohngo.ENERGY_LEFT));
				}
			}
		}
	}

	private void lflgfg() {
		if (currentAction != DFKLJDLKFJDYF) {
			Aummoo.play(LKGLKLDKF);
			rwrewe(DFKLJDLKFJDYF);
			aur.x = (int) x - 15;
			aur.y = (int) y - 50;
		} else {
			if (bughaet.getFfdklfjdyfame() == 4 && bughaet.kdfkdjklf() == 0) {
				for (int c = 0; c < 3; c++) {
					soohngos.add(
							new Soohngo(ohvzoob, aur.x + aur.width / 2, aur.y + 5, Soohngo.ENERGY_UP));
				}
			}
		}
	}

	@Override
	public void draw(Graphics2D g) {

		// draw emote
		if (jdkfd == FJKDLJFDF_DFDF) {
			g.drawImage(pofdof, (int) (x + xmap - cwidth / 2.0), (int) (y + ymap - 40), null);
		} else if (jdkfd == JDKFJDF_DFJDKF) {
			g.drawImage(kfldfd, (int) (x + xmap - cwidth / 2.0), (int) (y + ymap - 40), null);
		}

		// draw energy particles
		for (int i = 0; i < soohngos.size(); i++) {
			soohngos.get(i).draw(g);
		}

		// flinch
		if (opfgpoidfgidfg && !zetrer && wijeern % 10 < 5) {
			return;
		}

		super.draw(g);

	}
}